feat: add performance improvements and more test scenarios

+++ src/main/kotlin/io/github/luizimcpi/model/entity/Transacao.kt
@@ -18,7 +18,7 @@ data class Transacao(
     var id: Long? = null,
 
     @Column
-    val valor: Long,
+    val valor: Int,
 
     @Column
     val tipo: String,
